+++ LenovoLegionToolkit.WPF/Windows/Automation/TabItemContent/IAutomationPipelineTriggerTabItemContent.cs
@@ -1,0 +1,8 @@
+﻿using LenovoLegionToolkit.Lib.Automation.Pipeline.Triggers;
+
+namespace LenovoLegionToolkit.WPF.Windows.Automation.TabItemContent;
+
+public interface IAutomationPipelineTriggerTabItemContent<out T> where T : IAutomationPipelineTrigger
+{
+    T GetTrigger();
+}
