+++ LenovoLegionToolkit.WPF/Windows/Automation/TabItemContent/TimeTabItem.xaml
@@ -1,5 +1,5 @@
 ﻿<UserControl
-    x:Class="LenovoLegionToolkit.WPF.Windows.Automation.TabItem.TimeTabItem"
+    x:Class="LenovoLegionToolkit.WPF.Windows.Automation.TabItemContent.TimeTabItemContent"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
